check pipelinerun can serialize correctly in defaulter
When it comes to serialization, PipelineRuns carry a few semantics about
their data, including that parameters must have a value.  In practice,
however, we don't see these semantics enforced except for when
pipelineruns are attempting to be serialized.  This means we can run our
webhook successfully, but controller-runtime returns a 403 Forbidden
response to the user because their PipelineRun was bad.  In these cases,
instead return a 400 Bad Request error.

This is cropping up because defaulting webhooks run before validation
webhooks, and so we see errors before tekton's validation webhooks can
run.

Signed-off-by: Andy Sadler <ansadler@redhat.com>